+++ functionGenerator.h
@@ -5,7 +5,6 @@
 // VERSION: 0.2.5
 // PURPOSE: wave form generating functions (use with care)
 //     URL: https://github.com/RobTillaart/FunctionGenerator
-//
 
 
 #include "Arduino.h"